Update log_production_model.py
runs = mlflow.search_runs(experiment_ids=[1])

Sir, I've followed your code files and after replicating everything there is an error being raised at "log_production_model" stage, and the issue is resolved when the experiment id is given as list instead of an index. I don't know the reason for the error but it worked for you but is raising an issue in my case.